refactor: Section 일급 컬렉션 분리
